water_world/requirements.txt
    flask==3.0.2
    werkzeug==3.0.1
    flask-socketio==5.3.6
    flask-cors==4.0.0
    python-engineio==4.9.1
    python-socketio==5.11.3
    PyJWT==2.9.0
    requests==2.31.0
    gunicorn==21.2.0
    eventlet==0.39.1
    dnspython==2.4.2
    kafka-python==2.2.15